requirements.txt

# paddlepaddle-gpu==2.5.0.post118  # Commented out for CPU only setup
gradio==5.38.2
paddleocr==2.6.1.3
paddlepaddle==2.5.0
pdf2image==1.17.0
PyMuPDF==1.20.2
pytesseract==0.3.10
pillow==11.3.0
transformers==4.49.0
torch==2.7.1
spacy==3.8.7
spacy-transformers==1.3.9
pandas==1.5.3
opencv-python-headless==4.12.0.88
huggingface-hub==0.34.1


